fix: bugs found by widget tests
