+++ test/widget_test/details/details_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:network_image_mock/network_image_mock.dart';
-import 'package:projeto_criptos/details/widgets/details_screen.dart';
+import 'package:projeto_criptos/details/view/details_screen.dart';
 import 'package:projeto_criptos/shared/utils/decimal_parse.dart';
 
 import '../../setup/common_asset.dart';
